Bug fix when using "next":
we can't decide on how to step until the available disassembly matches
the PC we are stepping from. This slows down stepping a little (since there is
a round-trip). In theory we could send the next instruction opcode as part of
the "status" notification if we want it to be faster.