Hatari: attempted fix for losing CPU instructions/cycles when start/stop used.
Normally hatari called Profile_CpuStart and this wipes all previous
instruction info, so trying to accumulate start/stop can lose data.
This patch tries to save the previous instruction information even when
the profiling is not "active", so that when profiling starts the previous
instruction time is still included correctly.