Hatari: another attempt to work round the way profling works.
No part of the debug system updates until a debugger requests a break,
so any attempt to update the "previous instruction" will fail before it
is enabled.

The workaround is to check the "previous cycles" value is 0, and if so
to use the current cycles value. This is inaccurate (since we can break
partway through the previous instruction) but the best we can do given
the way Hatari works :(